+++ addons/actions/configure.js
@@ -2,7 +2,8 @@ import {configureActions as addonConfigureActions} from '@storybook/addon-action
 
 const configureActions = opts => {
 	return addonConfigureActions({
-		// TODO: Set to limit of 200 is a workaround. We need to update event display to sort in latest order.
+		// TODO: Set the limit of 200 as a workaround.
+		// We hope storybook to fix the action log to be sorted as the latest on the top.
 		// Limit the number of items logged into the actions panel
 		limit: 200,
 		...opts